move include files to include/clRNG
