+++ src/client/Inventory/InventoryKernels.cl
@@ -32,7 +32,7 @@
 
 #define CLRNG_ENABLE_SUBSTREAMS
 /*! [clRNG header] */
-#include <mrg31k3p.clh>
+#include <clRNG/mrg31k3p.clh>
 /*! [clRNG header] */
 
 __constant int L = 100;
